test(sdk-trace-web): run sdk tests on web worker
