+++ .github/workflows/unit-test.yml
@@ -81,6 +81,41 @@ jobs:
         run: npm run test:browser
       - name: Report Coverage
         run: npm run codecov:browser
+  webworker-tests-stable:
+    runs-on: ubuntu-latest
+    container:
+      image: circleci/node:14-browsers
+    env:
+      NPM_CONFIG_UNSAFE_PERM: true
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v1
+      - name: Permission Setup
+        run: sudo chmod -R 777 /github /__w
+
+      - name: restore lerna
+        uses: actions/cache@v2
+        id: cache
+        with:
+          path: |
+            node_modules
+            */*/node_modules
+          key: browser-tests-stable-${{ runner.os }}-${{ matrix.node_version }}-${{ hashFiles('**/package.json') }}
+
+      - name: Bootstrap
+        if: steps.cache.outputs.cache-hit != 'true'
+        run: |
+          npm install --ignore-scripts
+          npx lerna bootstrap --no-ci --hoist --nohoist='zone.js'
+
+      - name: Build 🔧
+        run: |
+          npm run compile
+
+      - name: Unit tests
+        run: npm run test:webworker
+      - name: Report Coverage
+        run: npm run codecov:webworker
   node-tests-experimental:
     strategy:
       fail-fast: false
@@ -164,3 +199,42 @@ jobs:
       - name: Report Coverage
         working-directory: experimental
         run: npm run codecov:browser
+  webworker-tests-experimental:
+    runs-on: ubuntu-latest
+    container:
+      image: circleci/node:14-browsers
+    env:
+      NPM_CONFIG_UNSAFE_PERM: true
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v1
+      - name: Permission Setup
+        run: sudo chmod -R 777 /github /__w
+
+      - name: restore lerna
+        uses: actions/cache@v2
+        id: cache
+        with:
+          path: |
+            experimental/node_modules
+            experimental/*/*/node_modules
+          key: browser-tests-experimental-${{ runner.os }}-${{ matrix.node_version }}-${{ hashFiles('**/package.json') }}
+
+      - name: Bootstrap
+        if: steps.cache.outputs.cache-hit != 'true'
+        working-directory: experimental
+        run: |
+          npm install --ignore-scripts
+          npx lerna bootstrap --no-ci --hoist --nohoist='zone.js'
+
+      - name: Build 🔧
+        working-directory: experimental
+        run: |
+          npm run compile
+
+      - name: Unit tests
+        working-directory: experimental
+        run: npm run test:webworker
+      - name: Report Coverage
+        working-directory: experimental
+        run: npm run codecov:webworker
