+++ package.json
@@ -16,11 +16,13 @@
     "version:update": "lerna run version:update",
     "test": "lerna run test",
     "test:browser": "lerna run test:browser",
+    "test:webworker": "lerna run test:webworker",
     "test:backcompat": "lerna run test:backcompat",
     "bootstrap": "lerna bootstrap --hoist --nohoist='zone.js'",
     "changelog": "lerna-changelog",
     "codecov": "lerna run codecov",
     "codecov:browser": "lerna run codecov:browser",
+    "codecov:webworker": "lerna run codecov:webworker",
     "predocs-test": "npm run docs",
     "docs": "typedoc && touch docs/.nojekyll",
     "docs-deploy": "gh-pages --dotfiles --dist docs",
