+++ packages/opentelemetry-sdk-trace-web/package.json
@@ -13,9 +13,11 @@
     "lint": "eslint . --ext .ts",
     "lint:fix": "eslint . --ext .ts --fix",
     "codecov:browser": "nyc report --reporter=json && codecov -f coverage/*.json -p ../../",
+    "codecov:webworker": "nyc report --reporter=json && codecov -f coverage/*.json -p ../../",
     "version": "node ../../scripts/version-update.js",
     "tdd": "karma start",
     "test:browser": "nyc karma start --single-run",
+    "test:webworker": "nyc karma start karma.worker.js --single-run",
     "watch": "tsc --build --watch tsconfig.all.json",
     "precompile": "lerna run version --scope $(npm pkg get name) --include-dependencies",
     "prewatch": "npm run precompile"
@@ -69,6 +71,7 @@
     "karma-coverage-istanbul-reporter": "3.0.3",
     "karma-jquery": "0.2.4",
     "karma-mocha": "2.0.1",
+    "karma-mocha-webworker": "1.3.0",
     "karma-spec-reporter": "0.0.32",
     "karma-webpack": "4.0.2",
     "mocha": "7.2.0",
